Split notebooks into separate files
