+++ tests/figure.spec.ts-snapshots/can-open-a-simple-figure-1-chromium-linux.png
