+++ tests/table.spec.ts-snapshots/can-open-a-simple-table-1-chromium-linux.png
